dataipa: Resolves makefile violations in dataipa
During the makefile violation scan for the Bazel migration on Android-V,
cnt_macro_errors were identified in the dataipa makefile. The current is-xxx macros have several issues, as they are awkwardly used by comparing their result with true instead of the conventional method of checking if the value is not empty. Therefore, these macros are being replaced with a new, more convenient set of macros defined in build/make/core.

Change-Id: Iad5729d7c9d3c511e4a903583c458594144283eb
Signed-off-by: Jagadeesh Ponduru <quic_jponduru@quicinc.com>